- Update logic server

+++ ServerGame/Server/Object.cpp
@@ -124,7 +124,7 @@ void Object::SetVel(float _x, float _y)
 	box->vy = _y;
 }
 
-void Object::SetBoxSize(int _width, int _height)
+void Object::SetBoxSize(float _width, float _height)
 {
 	box->width = _width;
 	box->height = _height;
